+++ python/lsst/ts/genericcamera/genericcameracsc.py
@@ -27,6 +27,9 @@ import traceback
 import inspect
 import os
 import logging
+import yaml
+
+import numpy as np
 
 from .config_schema import CONFIG_SCHEMA
 from . import __version__
@@ -39,6 +42,11 @@ from . import driver
 LV_ERROR = 1000
 """Error code for when the live view loop dies and the CSC is in enable
 state.
+"""
+
+AE_ERROR = 2000
+"""Error code for when the auto exposure loop dies and the CSC is in
+enable state.
 """
 
 
@@ -100,9 +108,12 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
         self.server = None
 
         self.isLive = False
+        self.isAutoExposure = False
         self.isExposing = False
         self.runLiveTask = False
         self.liveTask = None
+        self.runAutoExposureTask = False
+        self.autoExposureTask = None
         self.log.debug("Generic Camera CSC Ready")
 
     async def begin_enable(self, id_data):
@@ -173,15 +184,15 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
                 self.camera = None
 
     async def do_setValue(self, id_data):
-        """Sets a parameter/value pair.
+        """Set a parameter/value pair.
 
         Parameters
         ----------
         id_data :
-            id : int
+            id : `int`
                 The command id.
-            data : GenericCamera_command_setValueC
-                parametersAndValues : str
+            data : `GenericCamera_command_setValueC`
+                parametersAndValues : `str`
                     A comma deliminated pair key,value.
         """
         self.assert_enabled("setValue")
@@ -193,22 +204,23 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
         self.log.info("setValue - End")
 
     def do_setROI(self, id_data):
-        """Sets the region of interest.
+        """Set the region of interest.
 
         Parameters
         ----------
         id_data :
-            id : int
+            id : `int`
                 The command id.
-            data : GenericCamera_command_setROIC
-                topPixel : int
+            data : `GenericCamera_command_setROIC`
+                topPixel : `int`
                     The top pixel of the RIO in binned pixels.
-                leftPixel : int
+                leftPixel : `int`
                     The left pixel of the RIO in binned pixels.
-                width : int
+                width : `int`
                     The width of the ROI in binned pixels.
-                height : int
-                    The height of the ROI in binned pixels."""
+                height : `int`
+                    The height of the ROI in binned pixels.
+        """
         self.assert_enabled("setROI")
         self._assert_notlive()
 
@@ -228,16 +240,17 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
             self.log.warning("ROI already set with same parameters.")
 
     def do_setFullFrame(self, id_data):
-        """Sets the region of interest to full frame.
+        """Set the region of interest to full frame.
 
         Parameters
         ----------
         id_data :
-            id : int
+            id : `int`
                 The command id.
-            data : GenericCamera_command_setFullFrameC
+            data : `GenericCamera_command_setFullFrameC`
                 ignored : bool
-                    This is ignored."""
+                    This is ignored.
+        """
         self.log.info("setFullFrame - Start")
         self.assert_enabled("setFullFrame")
         self._assert_notlive()
@@ -258,6 +271,7 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
         self.assert_enabled("startLiveView")
         self.log.info("startLiveView - Start")
         self._assert_notlive()
+        self._assert_notautoexposure()
         if id_data.expTime == 0.0:
             raise RuntimeError("LiveView exposure time must be greater than zero.")
         self.camera.startLiveView()
@@ -268,16 +282,17 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
         self.log.info("startLiveView - End")
 
     async def do_stopLiveView(self, id_data):
-        """Stops the live view display.
+        """Stop the live view display.
 
         Parameters
         ----------
         id_data :
-            id : int
+            id : `int`
                 The command id.
-            data : GenericCamera_command_stopLiveViewC
-                ignored : bool
-                    This is ignored."""
+            data : `GenericCamera_command_stopLiveViewC`
+                ignored : `bool`
+                    This is ignored.
+        """
         self.assert_enabled("stopLiveView")
         self._assert_live()
 
@@ -308,23 +323,24 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
         self.evt_endLiveView.put()
 
     async def do_takeImages(self, id_data):
-        """Starts taking images.
+        """Start taking images.
 
         Parameters
         ----------
         id_data :
-            id : int
+            id : `int`
                 The command id.
-            data : GenericCamera_command_takeImagesC
-                numImages : int
+            data : `GenericCamera_command_takeImagesC`
+                numImages : `int`
                     The number of images to take in the sequence.
-                expTime : float
+                expTime : `float`
                     The exposure time in seconds.
-                shutter : bool
+                shutter : `bool`
                     True if the shutter should be utilized.
         """
         self.assert_enabled("takeImages")
         self._assert_notlive()
+        self._assert_notautoexposure()
         self.isExposing = True
 
         try:
@@ -350,46 +366,13 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
                     index=imageIndex,
                     total=imagesInSequence,
                 )
-                if id_data.shutter:
-                    self.evt_startShutterOpen.put()
-                    await self.camera.startShutterOpen()
-
-                    await self.camera.endShutterOpen()
-                    self.evt_endShutterOpen.put()
-
-                self.evt_startIntegration.set_put(
-                    imagesInSequence=imagesInSequence,
-                    imageName=imageName,
-                    imageIndex=imageIndex,
-                    timestampAcquisitionStart=timeStamp,
-                    exposureTime=exposureTime,
-                )
-                await self.camera.startIntegration()
-                await self.camera.endIntegration()
-                self.evt_endIntegration.put()
-
-                if id_data.shutter:
-                    self.evt_startShutterClose.put()
-                    await self.camera.startShutterClose()
-
-                    await self.camera.endShutterClose()
-                    self.evt_endShutterClose.put()
-                self.evt_startReadout.set_put(
-                    imagesInSequence=imagesInSequence,
-                    imageName=imageName,
-                    imageIndex=imageIndex,
-                    timestampAcquisitionStart=timeStamp,
-                    exposureTime=exposureTime,
-                )
-                await self.camera.startReadout()
-
-                exposure = await self.camera.endReadout()
-                self.evt_endReadout.set_put(
-                    imagesInSequence=imagesInSequence,
-                    imageName=imageName,
-                    imageIndex=imageIndex,
-                    timestampAcquisitionStart=timeStamp,
-                    requestedExposureTime=exposureTime,
+                exposure = await self.take_image(
+                    shutter=id_data.shutter,
+                    images_in_sequence=imagesInSequence,
+                    image_index=imageIndex,
+                    exposure_time=exposureTime,
+                    timestamp=timeStamp,
+                    image_name=imageName,
                 )
                 exposure.save(os.path.join(self.directory, imageName + ".fits"))
             await self.camera.endTakeImage()
@@ -401,12 +384,158 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
             self.isExposing = False
         self.log.info("takeImages - End")
 
+    async def take_image(
+        self,
+        shutter,
+        images_in_sequence,
+        image_index,
+        exposure_time,
+        timestamp,
+        image_name,
+    ):
+        """Take a single image with the given parameters.
+
+        Parameters
+        ----------
+        shutter: `bool`
+            True if the shutter should be utilized.
+        images_in_sequence: `int`
+            The number of images to take in the sequence.
+        image_index: `int`
+            The index of the image in the sequence.
+        exposure_time: `float`
+            The exposure time in seconds.
+        timestamp: `float`
+            The time in seconds.
+        image_name: `str`
+            The name of the image file.
+        Returns
+        -------
+        exposure: `exposure.Exposure`
+            The exposure.
+        """
+        if shutter:
+            self.evt_startShutterOpen.put()
+            await self.camera.startShutterOpen()
+
+            await self.camera.endShutterOpen()
+            self.evt_endShutterOpen.put()
+
+        self.evt_startIntegration.set_put(
+            imagesInSequence=images_in_sequence,
+            imageName=image_name,
+            imageIndex=image_index,
+            timestampAcquisitionStart=timestamp,
+            exposureTime=exposure_time,
+        )
+        await self.camera.startIntegration()
+        await self.camera.endIntegration()
+        self.evt_endIntegration.put()
+
+        if shutter:
+            self.evt_startShutterClose.put()
+            await self.camera.startShutterClose()
+
+            await self.camera.endShutterClose()
+            self.evt_endShutterClose.put()
+        self.evt_startReadout.set_put(
+            imagesInSequence=images_in_sequence,
+            imageName=image_name,
+            imageIndex=image_index,
+            timestampAcquisitionStart=timestamp,
+            exposureTime=exposure_time,
+        )
+        await self.camera.startReadout()
+
+        exposure = await self.camera.endReadout()
+        self.evt_endReadout.set_put(
+            imagesInSequence=images_in_sequence,
+            imageName=image_name,
+            imageIndex=image_index,
+            timestampAcquisitionStart=timestamp,
+            requestedExposureTime=exposure_time,
+        )
+        return exposure
+
+    async def do_startAutoExposure(self, id_data):
+        """Start taking exposures automatically.
+
+        Parameters
+        ----------
+        id_data :
+            minExpTime : `float`
+                The minimum exposure time in seconds.
+            maxExpTime : `float`
+                The maximum exposure time in seconds.
+            configuration : `str`
+                A Yaml string containing additional configuration
+                parameters.
+        """
+        self.assert_enabled("startAutoExposure")
+        self.log.info("startAutoExposure - Start")
+        self._assert_notlive()
+        self._assert_notautoexposure()
+        self.runAutoExposureTask = True
+
+        # Prepare potential configuration items for the loop
+        configuration = {
+            "shutter": False,
+            "sensors": "",
+            "keyValueMap": "",
+            "obsNote": "",
+        }
+        if id_data.configuration != "":
+            loaded_configuration = yaml.safe_load(id_data.configuration)
+            for key in configuration:
+                if key in loaded_configuration:
+                    configuration[key] = loaded_configuration[key]
+
+        self.autoExposureTask = asyncio.ensure_future(
+            self.run_auto_exposure_loop(
+                id_data.minExpTime, id_data.maxExpTime, configuration
+            )
+        )
+        self.evt_autoExposureStarted.set_put(
+            minExpTime=id_data.minExpTime,
+            maxExpTime=id_data.maxExpTime,
+            configuration=id_data.configuration,
+        )
+        self.log.info("startAutoExposure - End")
+
+    async def do_stopAutoExposure(self, id_data):
+        """Stop taking exposures automatically.
+
+        Parameters
+        ----------
+        id_data
+            Nothing passed on.
+        """
+        self.assert_enabled("stopAutoExposure")
+        self._assert_autoexposure()
+
+        self.log.info("stopAutoExposure - Start")
+
+        self.runAutoExposureTask = False
+        await self.autoExposureTask
+
+        await self.stop_autoexposure()
+
+        self.log.info("stopAutoExposure - End")
+
+    async def stop_autoexposure(self):
+        """Stop auto exposure.
+
+        Not much content for now but this may change in the future.
+        """
+
+        self.evt_autoExposureStopped.put()
+
     async def liveView_loop(self, exposure_time):
         """Run the live view capture loop.
 
         Parameters
         ----------
-        exposure_time : float
+        exposure_time : `float`
             The exposure time of the image (in seconds).
         """
         self.log.debug("liveView_loop - Start")
@@ -455,23 +584,340 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
         self.isLive = False
         self.log.info("liveView_loop - End")
 
+    async def run_auto_exposure_loop(self, min_exp_time, max_exp_time, configuration):
+        """Prepare and start the auto exposure capture loop.
+
+        The cadence of the images is determined by the value of
+        config.autoExposureInterval.
+
+        Parameters
+        ----------
+        min_exp_time: `float`
+            The minimum exposure time to use.
+        max_exp_time: `float`
+            The maximum exposure time to use.
+        configuration: `dict`
+            A dict containing additional configuration parameters.
+        """
+        self.log.info("autoExposure_loop - Start")
+        self.isAutoExposure = True
+
+        augmented_configuration = configuration.copy()
+        augmented_configuration["images_in_sequence"] = 0
+        augmented_configuration["image_index"] = 0
+        augmented_configuration["exposure_time"] = min_exp_time
+
+        try:
+            await self.run_auto_exposure(
+                min_exp_time, max_exp_time, augmented_configuration
+            )
+        except Exception:
+            self.log.exception("Error in auto exposure loop.")
+            await self.stop_autoexposure()
+
+            self.fault(
+                code=AE_ERROR,
+                report="Error in auto exposure loop.",
+                traceback=traceback.format_exc(),
+            )
+
+        self.isAutoExposure = False
+        self.log.info("autoExposure_loop - End")
+
+    async def run_auto_exposure(self, min_exp_time, max_exp_time, configuration):
+        """Take auto exposures until `runAutoExposureTask` is False.
+
+        Parameters
+        ----------
+        min_exp_time: `float`
+            The minimum exposure time to use.
+        max_exp_time: `float`
+            The maximum exposure time to use.
+        configuration: `dict`
+            A dict containing additional configuration parameters.
+        """
+
+        # First determine the exposure time based by taking images
+        # starting from the configured minimum exposure time.
+        timestamp = time.time()
+        image_name = self.fileNameFormat.format(
+            timestamp=int(timestamp), index=0, total=1
+        )
+        exposure_time_auto_current = await self.determine_exposure_time(
+            min_exp_time, max_exp_time, configuration, timestamp, image_name
+        )
+
+        self.log.debug(
+            f"Initial auto exposure time: {exposure_time_auto_current}s "
+            f"[{min_exp_time}:{max_exp_time}]"
+        )
+
+        # Then loop and take images using the latest exposure time and
+        # update the exposure time if necessary on the way.
+        while self.runAutoExposureTask:
+            timestamp = time.time()
+            image_name = self.fileNameFormat.format(
+                timestamp=int(timestamp),
+                index=configuration["image_index"],
+                total=configuration["images_in_sequence"],
+            )
+
+            self.evt_startTakeImage.put()
+            await self.camera.startTakeImage(
+                exposure_time_auto_current,
+                configuration["shutter"],
+                configuration["sensors"],
+                configuration["keyValueMap"],
+                configuration["obsNote"],
+            )
+
+            # Create a sleep task to wait for while taking the image.
+            task_timer = asyncio.create_task(
+                asyncio.sleep(self.config.autoExposureInterval)
+            )
+            # Take an exposure until the background level is within
+            # the expected boundaries. This updates the exposure time
+            # as well.
+            (
+                exposure,
+                exposure_time_auto_new,
+            ) = await self.get_auto_exposure_and_exposure_time(
+                min_exp_time,
+                max_exp_time,
+                configuration,
+                timestamp,
+                image_name,
+                exposure_time_auto_current,
+            )
+
+            if exposure:
+                # Save the image.
+                exposure.save(os.path.join(self.directory, image_name + ".fits"))
+                # Update the initial exposure time for the next run of
+                # the loop.
+                self.log.debug(
+                    f"Auto exposure time adjusted: {exposure_time_auto_current}s "
+                    f"-> {exposure_time_auto_new}s."
+                )
+                exposure_time_auto_current = exposure_time_auto_new
+
+            await self.camera.endTakeImage()
+            self.evt_endTakeImage.put()
+
+            # Now await the sleep task so either taking images is done
+            # or a new one gets scheduled.
+            await task_timer
+
+    async def determine_exposure_time(
+        self,
+        min_exp_time,
+        max_exp_time,
+        configuration,
+        timestamp,
+        image_name,
+    ):
+        """Take images starting with the configured minimum exposure
+        time and increasing it up to the maximum exposure time if
+        necessary while determining the background level and verify
+        it that is within the configured limits.
+
+        Parameters
+        ----------
+        min_exp_time: `float`
+            The minimum exposure time to use.
+        max_exp_time: `float`
+            The maximum exposure time to use.
+        configuration: `dict`
+            A dict containing configuration parameters.
+        timestamp: `float`
+            The timestamp of the exposure [s].
+        image_name: `str`
+            The name of the exposure file.
+
+        Returns
+        -------
+        exposure_time: `float`
+            The exposure time determined by inspecting the background
+            level of images taken with varying exposure times.
+        """
+        exposure_time = configuration["exposure_time"]
+
+        await self.camera.startTakeImage(
+            exposure_time,
+            configuration["shutter"],
+            configuration["sensors"],
+            configuration["keyValueMap"],
+            configuration["obsNote"],
+        )
+
+        background_level = 0.0
+        while not (
+            self.config.minBackground <= background_level <= self.config.maxBackground
+        ):
+            self.log.debug("Taking exposure.")
+            exposure = await self.take_image(
+                configuration["shutter"],
+                configuration["images_in_sequence"],
+                configuration["image_index"],
+                exposure_time,
+                timestamp,
+                image_name,
+            )
+            self.log.debug("Establishing exposure background level.")
+            background_level = self.establish_exposure_background(exposure)
+            self.log.debug(
+                f"Background level is {background_level} and a value between "
+                f"{self.config.minBackground} and {self.config.maxBackground} "
+                f"is expected."
+            )
+            new_exposure_time = await self.adjust_exposure_time(
+                min_exp_time,
+                max_exp_time,
+                configuration,
+                background_level,
+                exposure_time,
+            )
+
+            if (
+                not (
+                    self.config.minBackground
+                    <= background_level
+                    <= self.config.maxBackground
+                )
+                and new_exposure_time == exposure_time
+            ):
+                self.log.warn(
+                    "Cannot take an exposure with a valid background level. Ignoring."
+                )
+                break
+
+            exposure_time = new_exposure_time
+
+        return exposure_time
+
+    async def get_auto_exposure_and_exposure_time(
+        self,
+        min_exp_time,
+        max_exp_time,
+        configuration,
+        timestamp,
+        image_name,
+        initial_exposure_time,
+    ):
+        """Take exposures and adjust the exposure time if necessary
+        based on the background level of the image.
+
+        Parameters
+        ----------
+        min_exp_time: `float`
+            The minimum exposure time to use.
+        max_exp_time: `float`
+            The maximum exposure time to use.
+        configuration: `dict`
+            A dict containing additional configuration parameters.
+        timestamp: `float`
+            The timestamp of the exposure [s].
+        image_name: `str`
+            The name of the exposure file.
+        initial_exposure_time: `float`
+            The initial exposure time [s]
+
+        Returns
+        -------
+        exposure: `exposure.Exposure`
+            An exposure with a background level between the configured
+            minimum and maximim exposure times.
+        exposure_time: `float`
+            The exposure time to use by the next run of the loop.
+        """
+        exposure = None
+        exposure_time = initial_exposure_time
+
+        self.log.debug("Taking exposure.")
+        exposure = await self.take_image(
+            configuration["shutter"],
+            configuration["images_in_sequence"],
+            configuration["image_index"],
+            exposure_time,
+            timestamp,
+            image_name,
+        )
+        self.log.debug("Establishing exposure background level.")
+        background_level = self.establish_exposure_background(exposure)
+        self.log.debug(
+            f"Background level is {background_level} and a value between "
+            f"{self.config.minBackground} and {self.config.maxBackground} "
+            f"is expected."
+        )
+        new_exposure_time = await self.adjust_exposure_time(
+            min_exp_time,
+            max_exp_time,
+            configuration,
+            background_level,
+            exposure_time,
+        )
+
+        exposure_time = new_exposure_time
+        return exposure, exposure_time
+
+    async def adjust_exposure_time(
+        self, min_exp_time, max_exp_time, configuration, background_level, exposure_time
+    ):
+        """Adjust the exposure time based on the current exposure time
+        and the background level.
+
+        Parameters
+        ----------
+        min_exp_time: `float`
+            The minimum exposure time to use.
+        max_exp_time: `float`
+            The maximum exposure time to use.
+        configuration: `dict`
+            A dict containing additional configuration parameters.
+        background_level: `float`
+            The background level to compare against the background
+            level limits.
+        exposure_time: `float`
+            The current exposure time [s].
+
+        Returns
+        -------
+        new_exposure_time: `float`
+            The new exposure time based on the background level and
+            the current exposure time.
+        """
+        new_exposure_time = exposure_time
+        if background_level > self.config.maxBackground:
+            new_exposure_time = exposure_time / 2.0
+            if new_exposure_time < min_exp_time:
+                new_exposure_time = min_exp_time
+        elif background_level < self.config.minBackground:
+            new_exposure_time = exposure_time * 2.0
+            if new_exposure_time > max_exp_time:
+                new_exposure_time = max_exp_time
+        return new_exposure_time
+
+    def establish_exposure_background(self, exposure):
+        background_level = np.median(exposure.buffer)
+        return background_level
+
     @staticmethod
     def get_config_pkg():
         return "ts_config_ocs"
 
     async def configure(self, config):
-        """Implement method to configure the CSC.
+        """Configure the CSC.
 
         Parameters
         ----------
         config : `object`
             The configuration as described by the schema at ``schema_path``,
             as a struct-like object.
+
         Notes
         -----
         Called when running the ``start`` command, just before changing
         summary state from `State.STANDBY` to `State.DISABLED`.
-
         """
 
         self.ip = config.ip
@@ -489,10 +935,16 @@ class GenericCameraCsc(salobj.ConfigurableCsc):
 
     def _assert_notlive(self):
         """Raise an exception if live view is active."""
-        if self.isLive:
-            raise Exception()
+        assert not self.isLive, "Live view is active."
 
     def _assert_live(self):
         """Raise an exception if live view is not active."""
-        if not self.isLive:
-            raise Exception()
+        assert self.isLive, "Live view is not active."
+
+    def _assert_notautoexposure(self):
+        """Raise an exception if auto exposure is active."""
+        assert not self.isAutoExposure, "Auto exposure is active."
+
+    def _assert_autoexposure(self):
+        """Raise an exception if auto exposure is not active."""
+        assert self.isAutoExposure, "Auto exposure is not active."